requirements.txt
customtkinter==5.2.2
pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.24.3
easyocr==1.7.1
requests==2.31.0
python-supabase==2.0.0
aiohttp==3.9.3
asyncio==3.4.3 